+++ packages/frontend/src/components/feedback/radio-button.stories.ts
@@ -1,0 +1,25 @@
+import type { Meta, StoryObj } from '@storybook/react'
+
+import RadioButton from '@/components/feedback/radio-button'
+
+const meta: Meta<typeof RadioButton> = {
+  title: 'Feedback/Radio Button',
+  component: RadioButton,
+  parameters: {
+    backgrounds: {
+      default: 'White',
+    },
+  },
+}
+export default meta
+
+type Story = StoryObj<typeof RadioButton>
+
+export const Basic: Story = {
+  args: {
+    checked: true,
+    disabled: false,
+    name: 'Chrome',
+    value: 'chrome',
+  },
+}