+++ packages/frontend/src/components/speech/speech-mobile-toolbar.tsx
@@ -1,7 +1,6 @@
 'use client'
 import React, { useState, createContext, useContext, useEffect } from 'react'
 import styled from 'styled-components'
-import Link from 'next/link'
 // hook
 import useOutsideClick from '@/hooks/use-outside-click'
 // @twreporter
@@ -25,6 +24,8 @@ import TabBarButton from '@/components/button/tab-bar-button'
 import toastr from '@/utils/toastr'
 // constants
 import { Direction } from '@/components/speech'
+// util
+import { openFeedback } from '@/utils/feedback'
 
 // styles
 const MobileToolbarContainer = styled.div<{
@@ -93,14 +94,12 @@ const MobileToolbarContext = createContext<{
 const FeedbackButton: React.FC = () => {
   const { hideText } = useContext(MobileToolbarContext)
   return (
-    <ButtonContainer>
-      <Link href="/feedback">
-        <IconWithTextButton
-          text="問題回報"
-          iconComponent={<Report releaseBranch={releaseBranch} />}
-          hideText={hideText}
-        />
-      </Link>
+    <ButtonContainer onClick={openFeedback}>
+      <IconWithTextButton
+        text="問題回報"
+        iconComponent={<Report releaseBranch={releaseBranch} />}
+        hideText={hideText}
+      />
     </ButtonContainer>
   )
 }